+++ fuel/src/test/kotlin/com/github/kittinunf/fuel/core/requests/UploadRequestTest.kt
@@ -2,7 +2,6 @@ package com.github.kittinunf.fuel.core.requests
 
 import com.github.kittinunf.fuel.core.Blob
 import com.github.kittinunf.fuel.core.DataPart
-import com.github.kittinunf.fuel.core.DefaultRequest
 import com.github.kittinunf.fuel.core.FuelManager
 import com.github.kittinunf.fuel.core.Headers
 import com.github.kittinunf.fuel.core.Method
